feat: Moved the outpost and ship data to the data folder
