+++ src/bsgo/data/OutpostData.cc
@@ -1,0 +1,11 @@
+
+#include "OutpostData.hh"
+
+namespace bsgo {
+
+bool OutpostData::operator==(const OutpostData &rhs) const
+{
+  return dbId == rhs.dbId;
+}
+
+} // namespace bsgo
