+++ src/bsgo/data/ShipData.cc
@@ -1,0 +1,11 @@
+
+#include "ShipData.hh"
+
+namespace bsgo {
+
+bool ShipData::operator==(const ShipData &rhs) const
+{
+  return dbId == rhs.dbId;
+}
+
+} // namespace bsgo
