+++ src/bsgo/messages/loading/CMakeLists.txt
@@ -8,11 +8,9 @@ target_sources (bsgo_lib PRIVATE
 	${CMAKE_CURRENT_SOURCE_DIR}/DataSerialization.cc
 	${CMAKE_CURRENT_SOURCE_DIR}/LoadingFinishedMessage.cc
 	${CMAKE_CURRENT_SOURCE_DIR}/LoadingStartedMessage.cc
-	${CMAKE_CURRENT_SOURCE_DIR}/OutpostData.cc
 	${CMAKE_CURRENT_SOURCE_DIR}/OutpostListMessage.cc
 	${CMAKE_CURRENT_SOURCE_DIR}/PlayerListMessage.cc
 	${CMAKE_CURRENT_SOURCE_DIR}/PlayerLoginDataMessage.cc
-	${CMAKE_CURRENT_SOURCE_DIR}/ShipData.cc
 	${CMAKE_CURRENT_SOURCE_DIR}/ShipListMessage.cc
 	)
