+++ src/bsgo/messages/loading/OutpostListMessage.cc
@@ -1,5 +1,6 @@
 
 #include "OutpostListMessage.hh"
+#include "DataSerialization.hh"
 #include "SerializationUtils.hh"
 
 namespace bsgo {
