+++ tests/unit/bsgo/messages/loading/OutpostDataTest.cc
@@ -1,5 +1,5 @@
 
-#include "OutpostData.hh"
+#include "DataSerialization.hh"
 #include <gtest/gtest.h>
 
 using namespace ::testing;
