+++ tests/unit/bsgo/messages/loading/ShipDataTest.cc
@@ -1,5 +1,5 @@
 
-#include "ShipData.hh"
+#include "DataSerialization.hh"
 #include <gtest/gtest.h>
 
 using namespace ::testing;
